Automated releases
Set up an automated release flow in CI based on package versions. This requires that PRs that make changes to source code check in version deferrals with `yarn version check -i`. When those deferred versions are applied, we will automatically tag the release in GitLab, which will trigger the publish flows, as well as publishing a GitLab release with change logs.

Changelog: added