Add from and to commits to fix changelog generation
We don't use strictly semver-conforming tags, because this is a monorepo. In order to work around this, we need to pass the semver part of the tag, and then provide a specific range of commits, since GitLab can't look at tags anymore. So we now grab the commit of the last release of that package, and use a range from that commit to the current (tagged) commit.