Detect the first sentence that's completely on the page
This updates the mobile apps' behavior when detecting the first sentence on a page.

Previously, the apps would count the first sentence with any portion on the page as the first sentence. This led to a number of unexpected behaviors, such as flipping to the previous page when beginning to play audio.

Now, the apps only look at sentences whose contents are _completely_ on the page when determining which sentence is the first.

Changelog: changed