Fix index out of bounds error on Android
When listening to audio on the very last chapter of a book on Android, there was a bug that crashed the app. This is because getFragment always tried to figure out the progression through the current chapter by looking at the progression at the start of the next chapter, and wasn't accounting for when there was no next chapter!

Changelog: fixed