Support syncing positions across client devices
Closes #13

Adds a new API endpoint for storing and retrieving synced positions.

Updates the mobile apps to periodically sync to and fetch from this new endpoint. When in the foreground, the app will try to write the current position for each book to the backend every ten seconds. When in the background, it will use background fetch tasks to attempt to write the current position every fifteen minutes. If it finds a newer position (ordered by time, not progress through the book), it will update its local state.

Changelog: added