Split epub and fs utilities into their own packages
We'd like to be able to publish our epub library as a standalone package! We're starting by pulling it out into its own monorepo workspace so that it can be developed independently. This required pulling the `fs` util (`streamFile`) out as well, so that it could be used in both the new `@smoores/epub` and `@storyteller/web`.

Note on package names:

The eventual goal is to publish any publishable packages under the `storyteller` org (e.g. `@storyteller/epub`, `@storyteller/fs`, `@storyteller/cli`). At the moment, this is being (non-maliciously, I think) [squatted on in npm](https://npmjs.com/org/storyteller). I'm attempting to reach out to the current owner (who doesn't seem very active on npm or GitHub anymore), and have filed a ticket with npm/GitHub to get ownership of the org. I think this will eventually happen (I think we have a good case!), but in the meantime, we'll have to publish these under my personal npm org, as `@smoores/epub`, etc.

Changelog: other