Remove OpenBLAS option
OpenBLAS was added as a whisper.cpp build option a while ago, in the hopes that it would be somewhat faster for CPU-based transcription. This hasn't borne out -- OpenBLAS is the same or slightly slower than the standard CPU build without optimized models, which Storyteller doesn't have support for. Now that we have support for both AMD and NVIDIA GPU acceleration frameworks, it doesn't seem worth it to keep around this CPU based framework that doesn't provide any value for users.

Changelog: removed