Re-write UI with Mantine and Tailwind CSS
Until this point, the UI had been written with @ariakit/react, and I'd been managing the design entirely on my own. This is challenging for two reasons: first, I'm not a designer, and second, it's just too much code to write for what, ultimately, is a pretty simple UI.

So I've decided to move to [Mantine](https://mantine.dev), customizing with [Tailwind CSS](https://tailwindcss.com/) as needed. This is going to make it much easier to iterate on the web frontend!

While I was in here, I fixed a few other issues:

1. Added support for updating the authors and cover image in the actual resulting epub
2. Fixed #174 
3. Added support for updating user permissions after the user has been created
4. Properly fixed #175
5. Add a few usability features to the add book form

Changelog: changed