Handle replacing metadata when there are text nodes in the metadata element
As noted in !106, there was a bug in the `replaceMetadata` method. It assumed that the indexes into the array returned by `epub.getMetadata()` matched the indexes in the `<metadata>` element's children array, but this is not guaranteed to be true if there are non-metadata elements, such as text nodes, in this array. This isn't technically allowed by the spec, and is unlikely to exist, but it's better to handle these edge cases than have subtly incorrect behavior!

Changelog: fixed