Fix intersection observers
This both fixes the intersection observers for double tap listeners and moves back to the scanning fragment identification method.

The initial intersection observer approach was running twice, once for 0% intersecting and once for 100% intersecting. This meant that double tap observers were being added and then removed again for sentences that were partially on the page.

It also caused issues as an approach for correctly identifying the first visible fragment. Luckily, that code is actually quite fast, and so it didn't make much of a different to move those checks back out of the intersection observer.

Overall, we now emit the correct locator on every kind of navigation change much more robustly!

Changelog-mobile: fixed