Save and use words.txt from filesystem
I noticed that the backend was downloading the word list every time it processed a book. I think its better to fetch it once and save it to the filesystem so on subsequent runs we can just read it from the filesystem.

Additionally, we can load the Docker image with the word list already predownloaded.

Another benefit of this is that you can now run Storyteller without internet access (assuming of course you use something like Speaches for the transcribing).

Changelog-web: fixed