Improve cross-device position syncing.
On some devices such as Boox, the onLocatorChanged event is firing when the app returns to the foreground, which breaks syncing.

Unmount the Epub component when the app goes to the background to make syncing position between devices more reliable.

Also, in the Epub component, ignore locator changes if the location is deeply equal to the current one.

On Android, I've also observed thrashing of the locator when the Epub component mounts. To prevent this causing the locator timestamp to update erroneously, ignore locator changes for 1 second after mounting.

Changelog-mobile: fixed