[LMF]: Allow automatically importing files from folders
Adds a root level and collection level setting for automatic import paths. When set, Storyteller will use fs.watch to watch for changes to those folders (recursively) and re-scan when changes are detected. Scans are debounced such that they only run after five seconds of quiet, so as to attempt to avoid unnecessary rescans when multiple sequential changes are made to a library.

Changelog-web: added