Add support for Boox hardware page-turn button(s)
This adds a new plugin that modifies the MainActivity.kt to include a custom dispatchKeyEvent implementation that uses `DeviceEventManagerModule.RCTDeviceEventEmitter` to emit events into the JS runtime. *Note*: This makes use of `reactNativeHost.reactInstanceManager.currentReactContext` to obtain a reference to the DeviceEventEmitter module. Technically `getCurrentReactContext()` is meant to be private and is only visible for testing, but several other applications do this and the alternatives are much more verbose and complex. If this ever becomes actually private, we won't be screwed or anything, but we will probably have to look into broadcasting events in a way that they can be consumed by an Expo Module and emitted with Expo APIs.

Then, we add an event listener in the `<Epub>` component that calls goForward or goBackward depending on the keyCode pressed.

The keyCodes are configurable in the plugin, but they need to be known at build time, so if we've missed any (these should work for the Palma, Page, and Color, at least), we have to make a new build.

Changelog-mobile: added