fix: make it possible to build sqlite on macos
Running the instrumentation hook fails on MacOS because it expects the dynamic library loaded with `load_extension` to be in `sqlite/uuid.c.dylib`, rather than `sqlite/uuid.c.so`. 

I changed the build script to put it in `sqlite/uuid.c.dylib` instead on MacOS.

This could also be fixed by making the `UUID_EXT_PATH` here explicitly `uuid.c.so` instead, but that may cause issues on other platforms, not sure!

https://gitlab.com/storyteller-platform/storyteller/-/blob/a7dea65e83a2fd9cfa980d96db88949299d9d820/web/src/database/connection.ts#L24

Tangentially, if not running through `nix`, the default MacOS installation of `sqlite` will fail to compile. 
You need to install it through `homebrew` and do `C_INCLUDE_PATH=/opt/homebrew/opt/sqlite/include yarn dev:web` to make it work.

I think the `nix` route is more robust, so I didn't add this to the docs, but let me know if i should!