Clean up upload route
Pulls out the common functionality from autoimport and uploading. We now have core functions for:

- Reading metadata from both EPUBs and Audiobooks
- Writing metadata to both EPUBs and Audiobooks
- Determining when to override or merge metadata with the existing book record

This simplifies the code quite a bit and makes it _much_ more consistent.

Also, I broke up the upload route into a few different handler function for the different cases, to make that a bit easier to follow.