Add various lint rules for organizing imports
This basically changes every file in the codebase, but only by organizing the import rules. Here's what's new:

- Sort all import rules by group (builtin, external, workspace (e.g. `@storyteller-platform/epub`), path alias (e.g. `@/assets`), parent, sibling, index
- Sort imports alphabetically within groups
- Sort names alphabetically with imports
- Enforce `type` specifier on type-only imports

Also, I've renamed the workspace scope from `@storyteller`/`@smoores` to `@storyteller-platform`. It's not my favorite (I'd really prefer `@storyteller`), but it's not taken on npm, so I think we're stuck with it. I'll deprecate the old npm packages and publish them under the new scope.

One thing that's really annoying is that `@dword-design/import-alias` is _supposed_ to have an option to force sibling paths to use aliases, which is enabled, but it isn't working. It's possible that it's fixed in v6, but that seems to require flat config, which I don't want to deal with right now.