Fix message channel comms
I completely forgot about the progress updates when I first wrote this. I assumed that all of the updateBook calls would be sequential, but they're not at all. I used `port.once` to listen for responses from the main thread, which means that it's possible to end up in a situation where all of the listeners remove themselves without actually resolving, and the worker thread ends up hanging forever.

Changelog-web: fixed