Create initial book to statuses
Storyteller assumes that there's a book_to_status record for every book for every user. It ensures this by creating one for every user any time a new book is created.

But previously, it was not creating a new one for every book any time a new _user_ was created. This is now fixed.

This also includes a migration to backfill any missing book_to_statuses.

Changelog-web: fixed