Use react-window to virtualize book grid
For large libraries, actually rendering a grid with every book can be very expensive, especially during bulk actions, which can trigger re-mounts and re-renders.

This change introduces react-window, which allows us to only render a few rows of books at a time, while appearing to have one big list of books. It was kind of challenging to do this for two reasons:

1. This library isn't _really_ meant for flex-wrapped grid lists like we use, but we work around that by calculating the number of columns on resize
2. It's kind of challenging to have a dynamically sized header with a sticky toolbar beneath it like we do, but we work around _that_ by passing a ref to the header and measuring its size when it changes. Then we give space for the toolbar, and stickily position the toolbar in that space.

Changelog-web: improved