fix: don't hide search when no results are returned
Currently if you enter something in the search bar that yields no results on the books page, you won't be able to fix it as the search bar is hidden.

with search "aisohtnsiohtnasiotn"

Before:
![image](/uploads/dc51b912b5bce2148c391a6c78f7bdc4/image.png)

After:

![image](/uploads/da472f43c8969eccfa198a4143efc68c/image.png)

I did have a weird problem with an error occurring when unfocusing from the tab, but I think that's fixed by manually setting the rowcount to 0.

Changelog-web: fixed