Support ID3v2.3 tags
The audiobook library was previously using ffmpeg's default ID3 version, which is ID3v2.4, to write MP3 metadata tags. This actually corrupted ID3 tags for input files that used ID3v2.3, because ID3 tag names have specific lengths, and they differ between 2.3 and 2.4.

Now, we inspect the file to determine which ID3 version to use, and make sure to use that version when writing the tags, as well.

Changelog-web: fixed
Changelog-audiobook: fixed