feat: overhaul miniplayer once again!
This changes the default view of the miniplayer to be a small circle on the bottom right that can be dragged around.
When clicked it expands to the big miniplayer, which can also be positioned and pinned.

![Screen_Recording_2025-10-25_at_00.03.02](/uploads/d567b72a50ad06f8355b1fd50a8bd7e3/Screen_Recording_2025-10-25_at_00.03.02.mov)

In addition, change the chapter markers in the progress bar to be more like YouTubes, rather than have those weird circles

Before:
![image](/uploads/41e2778f4a509e0796908f9d81d2a536/image.png)


After:

![image](/uploads/e945b951ae90f9515f34e9b7c0442251/image.png)

Changelog-web: changed