fix: fix transcription continuing for every track on fail
How to produce the og error:
- Start a transcription of a book using the OpenAI/`whisper-server` method
- Let it transcribe a file or two (important that this should be after splitting the tracks)
- Kill the whisper server or otherwise cause an incorrect transcription

Note that your main server's memory usage does not seem to go down, and that you still occasionally get `Transcribing....` logs.

This is bc all the individual transcriptions are still being processed within the `Promise.all`. Especially for the `OpenAI` setting, the retry time for failed calls to the server is very long, on the order of 3ish minutes. 

This can cause problems on lower end hardware, by keeping a lot of the resources in memory, and by continuing to use a lot of resources bc of eg starting a new transcription.

This MR fixes it by manually tracking whether the transcription process has failed and ending each callback early if a failure is detected.

Changelog-web: fixed