Split up eslint config into per-package config
Ultimately, this was all to work around the fact that the @dword-design/eslint-plugin-import-alias just doesn't work very well with monorepos. Previously, we were able to run it in the CLI and on VS Code (somehow), but no other editors handled it correctly.

Also, one giant eslint file is kind of a pain to work with.

This is like a little non-standard because I wanted more control and flexibility than ESLint's (honestly kind of annoying) config extension/plugin system allowed, so the `@storyteller-platform/eslint` package just exports some plain objects that get spread into legacy configs. We will upgrade to flat config at some point (it'll hopefully be easier now with separate eslint configs).

Every config needs to be marked as a root, otherwise ESLint just keeps looking upwards ("config cascade") and finds the root file, which ignores everything.

If I never have to configure ESLint again it'll be to soon.

Changelog-web: maintenance