Prevent large font sizes from breaking the app
Limit font scaling and fix details about the UI
to prevent the app from breaking when font sizes are turned
up.

Later on, we can improve accessibility when font sizes are turned
up, but it's most important to just not break the UI when font sizes
are turned up.

Changelog-mobile: fixed